Updating log functions based on logging class pre-existing functionality. Separating tests into files based on new file structure.
